requirements.txt
gradio==3.32.0
numpy
pillow
onnxruntime
huggingface_hub
scikit-image
pandas
opencv-python>=4.6.0
requests
setuptools
pyperclip